backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
openai>=2.0,<3
python-dotenv>=1.0,<2
